+++ paper/src/main/java/me/gamerduck/alwaysauth/paper/PaperPlatform.java
@@ -2,19 +2,36 @@ package me.gamerduck.alwaysauth.paper;
 
 import io.papermc.paper.command.brigadier.CommandSourceStack;
 import me.gamerduck.alwaysauth.Platform;
+import me.gamerduck.alwaysauth.api.updates.ModrinthUpdateChecker;
 import me.gamerduck.alwaysauth.reflection.AuthenticationURLReplacer;
 import me.gamerduck.alwaysauth.reflection.ServerPropertiesReplacer;
+import io.papermc.paper.advancement.AdvancementDisplay;
+import net.kyori.adventure.text.Component;
+import net.kyori.adventure.text.format.NamedTextColor;
+import net.kyori.adventure.text.format.TextDecoration;
+import org.bukkit.*;
+import org.bukkit.advancement.Advancement;
+import org.bukkit.event.EventHandler;
+import org.bukkit.event.EventPriority;
+import org.bukkit.event.Listener;
+import org.bukkit.event.player.PlayerAdvancementDoneEvent;
+import org.bukkit.event.player.PlayerJoinEvent;
 import org.bukkit.plugin.java.JavaPlugin;
 
 import java.util.List;
 import java.util.logging.Logger;
 
-public class PaperPlatform extends Platform<CommandSourceStack> {
+public class PaperPlatform extends Platform<CommandSourceStack> implements Listener {
 
     private static final Logger LOGGER = Logger.getLogger("AlwaysAuth");
+    private final JavaPlugin plugin;
 
     public PaperPlatform(JavaPlugin bootstrap) {
         super(bootstrap.getDataFolder().toPath());
+        this.plugin = bootstrap;
+
+        String message = getUpdateMessage();
+        if (message != null) sendLogMessage(message);
 
         AuthenticationURLReplacer.replaceSessionService(this, config().getSessionServerUrl());
         ServerPropertiesReplacer.forcePreventProxyConnections(this);
@@ -48,8 +65,8 @@ public class PaperPlatform extends Platform<CommandSourceStack> {
             }
         });
 
-
     }
+
 
     @Override
     public void sendMessage(CommandSourceStack commandSender, String msg) {
@@ -69,6 +86,14 @@ public class PaperPlatform extends Platform<CommandSourceStack> {
     @Override
     public void sendWarningLogMessage(String msg) {
         LOGGER.warning(msg.replaceAll("§.", ""));
+    }
+
+    @EventHandler
+    public void onJoin(PlayerJoinEvent event) {
+        if (event.getPlayer().hasPermission("alwaysauth.admin")) {
+            String message = getUpdateMessage();
+            if (message != null) event.getPlayer().sendMessage(message);
+        }
     }
 
 }