+++ standalone/src/main/java/me/gamerduck/alwaysauth/StandalonePlatform.java
@@ -7,6 +7,8 @@ public class StandalonePlatform extends Platform<PrintStream>{
 
     public StandalonePlatform(Path dataDirectory) {
         super(dataDirectory);
+        String message = getUpdateMessage();
+        if (message != null) sendLogMessage(message);
     }
 
     @Override