+++ velocity/src/main/java/me/gamerduck/alwaysauth/velocity/VelocityPlatform.java
@@ -17,6 +17,8 @@ public class VelocityPlatform extends Platform<CommandSource> {
     public VelocityPlatform(ProxyServer server, Logger logger, Path dataDirectory) {
         super(dataDirectory);
         this.server = server;
+        String message = getUpdateMessage();
+        if (message != null) sendLogMessage(message);
 
         VelocitySessionServerChanger.setCustomSessionServer(this, config().getSessionServerUrl());
         VelocityConfigurationChanger.forcePreventClientProxyConnections(this, server);